2007-07-18 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* src/services/wfut/WfutSession.h, src/services/wfut/WfutSession.cpp, src/services/wfut/WfutService.h, src/services/wfut/WfutService.cpp: use size_t where needed

+++ src/services/wfut/WfutService.cpp
@@ -95,7 +95,7 @@ void WfutService::wfutSession_ServerListDownloading(const std::string &url)
 	DownloadingServerList.emit(url);
 }
 
-void WfutService::wfutSession_UpdatesCalculated(unsigned int numberOfFilesToUpdate)
+void WfutService::wfutSession_UpdatesCalculated(size_t numberOfFilesToUpdate)
 {
 	S_LOG_INFO("Wfut needs to download " << numberOfFilesToUpdate <<" updates.");
 	UpdatesCalculated.emit(numberOfFilesToUpdate);
