+++ src/services/wfut/WfutSession.h
@@ -37,7 +37,7 @@ public:
     WfutSession(sigc::slot<void, const std::string&, const std::string&>& downloadCompleteSlot 
     , sigc::slot<void, const std::string&, const std::string&, const std::string&>& downloadFailedSlot 
     , sigc::slot<void, const std::string&>& serverListDownloadingSlot
-    , sigc::slot<void, unsigned int>& updatesCalculatedSlot);
+    , sigc::slot<void, size_t>& updatesCalculatedSlot);
 
     virtual ~WfutSession();
